backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==1.10.13
python-dotenv==1.0.1
openai==1.14.2
pdfplumber==0.11.0
python-docx==1.1.0
python-pptx==0.6.23
